Refactored packages and directories.
