+++ src/main/TabMenuController.java
@@ -5,9 +5,10 @@
  */
 package main;
 
+import javafx.fxml.Initializable;
+
 import java.net.URL;
 import java.util.ResourceBundle;
-import javafx.fxml.Initializable;
 
 /**
  * FXML Controller class
@@ -17,15 +18,16 @@ import javafx.fxml.Initializable;
 public class TabMenuController implements Initializable {
 
     private Main mainApp;
+
     /**
      * Initializes the controller class.
      */
     @Override
     public void initialize(URL url, ResourceBundle rb) {
         // TODO
-    }    
-    
-    public void setMainApp(Main mainApp){
+    }
+
+    public void setMainApp(Main mainApp) {
         this.mainApp = mainApp;
     }
 }
