+++ src/main/view/ActivityPane.fxml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.layout.AnchorPane?>
+<AnchorPane xmlns:fx="http://javafx.com/fxml/1" id="AnchorPane" prefHeight="200.0" prefWidth="250.0"
+            xmlns="http://javafx.com/javafx/8" fx:controller="main.ActivityPaneController">
+    <children></children>
+</AnchorPane>
